Rename session_items
